second-commit

+++ animal/views.py
@@ -1,7 +1,7 @@
+from django.core.exceptions import ObjectDoesNotExist
 from django.shortcuts import render, get_object_or_404
 from django.http import HttpResponse
 from django.views.generic import ListView
-
 from .models import Animal
 
 
@@ -15,10 +15,16 @@ def create_animal(request):
     return render(request, 'create_animal.html')
 
 
-def delete_animal(request, animal_id):
-    animal = get_object_or_404(Animal, pk=animal_id)
-    animal.delete()
-    return HttpResponse(f'Animal {animal.name} deleted successfully!')
+def delete_animal(request):
+    if request.method == 'POST':
+        try:
+            animal_id = request.POST['animal_id']
+            animal = Animal.objects.get(pk=animal_id)
+        except ObjectDoesNotExist:
+            return HttpResponse('Animal does not exist, please check animal list')
+        animal.delete()
+        return HttpResponse(f'Animal with ID {animal_id} deleted successfully!')
+    return render(request, 'delete_animal.html')
 
 
 class AnimalListView(ListView):
